Rewrite most of the mod. This removes: boss, combat and dungeon music. This adds: shuffling playlists for each music type, better /music command, performance fixes, /biometag cmd.
