+++ gradle.properties
@@ -1,7 +1,7 @@
-mod_id=musify
-mod_name=Musify!
-mod_version=1.2.4
-mod_author=["Kwember"]
+mod_id=musicevents
+mod_name=MusicEvents
+mod_version=1.0.0
+mod_author=["Vortron-rd,Kwember"]
 mod_description=Customizable music tracks!
 mod_url="something"
