--- src/main/java/musify/musifyrewrite.txt
@@ -1,10 +1,0 @@
-musify stuff
-
-1 instantiable music player.
-this music player plays mp3.
-
-we have condition checks -> creates an instance of the player with a file selected from config.
-	-> saves the player in its own class to know what player to edit or change and saves it
-	   in the overarching list of active players
-
-all player instances in their own thread.
