--- src/main/java/musify/proxy/ClientProxy.java
@@ -1,20 +1,0 @@
-package musify.proxy;
-
-import musify.handlers.MainMenuMusicHandler;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-
-
-@SideOnly(Side.CLIENT)
-public class ClientProxy extends CommonProxy {
-
-    @Override
-    public void preInit(FMLPreInitializationEvent e) {
-        super.preInit(e);
-        MinecraftForge.EVENT_BUS.register(new MainMenuMusicHandler());
-    }
-
-
-}
