--- src/main/java/musify/proxy/CommonProxy.java
@@ -1,25 +1,0 @@
-package musify.proxy;
-
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-
-
-@Mod.EventBusSubscriber
-public class CommonProxy {
-
-    public void preInit(FMLPreInitializationEvent event) {
-
-    }
-
-    public void init(FMLInitializationEvent event) {
-
-    }
-
-    public void postInit(FMLPostInitializationEvent event) {
-
-
-    }
-
-}
