--- src/main/resources/META-INF/biomemusic_at.cfg
@@ -1,2 +1,0 @@
-# Allow access to the currentMusic field in MusicTicker
-public-f net.minecraft.client.audio.MusicTicker currentMusic
