+++ src/main/resources/mcmod.info
@@ -5,7 +5,7 @@
   "description": "${mod_description}",
   "version": "${version}",
   "mcversion": "${mcversion}",
-  "url": "https://legacy.curseforge.com/minecraft/mc-mods/musify",
+  "url": "",
   "updateUrl": "",
   "authorList": ${mod_author},
   "credits": "",
