+++ src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
   "pack": {
-    "description": "Musify Assets",
+    "description": "musicevents Assets",
     "pack_format": 3
   }
 }
